work item icons and split of files
