+++ package.json
@@ -1,53 +1,53 @@
 {
-    "name": "azure-boards-vs-code",
-    "displayName": "azure-boards-vs-code",
-    "description": "",
-    "version": "0.0.1",
-    "publisher": "cschleiden",
-    "engines": {
-        "vscode": "^1.27.0"
-    },
-    "categories": [
-        "Other"
-    ],
-    "activationEvents": [
-        "onView:work-items"
-    ],
-    "main": "./out/extension",
-    "contributes": {
-        "viewsContainers": {
-            "activitybar": [
-                {
-                    "id": "azure-boards-container",
-                    "title": "Azure Boards",
-                    "icon": "media/logo.svg"
-                }
-            ]
-        },
-        "views": {
-            "azure-boards-container": [
-                {
-                    "id": "work-items",
-                    "name": "My work items"
-                }
-            ]
+  "name": "azure-boards-vs-code",
+  "displayName": "azure-boards-vs-code",
+  "description": "",
+  "version": "0.0.1",
+  "publisher": "cschleiden",
+  "engines": {
+    "vscode": "^1.27.0"
+  },
+  "categories": [
+    "Other"
+  ],
+  "activationEvents": [
+    "onView:work-items"
+  ],
+  "main": "./out/extension",
+  "contributes": {
+    "viewsContainers": {
+      "activitybar": [
+        {
+          "id": "azure-boards-container",
+          "title": "Azure Boards",
+          "icon": "media/logo.svg"
         }
+      ]
     },
-    "scripts": {
-        "vscode:prepublish": "npm run compile",
-        "compile": "tsc -p ./",
-        "watch": "tsc -watch -p ./",
-        "postinstall": "node ./node_modules/vscode/bin/install",
-        "test": "npm run compile && node ./node_modules/vscode/bin/test"
-    },
-    "devDependencies": {
-        "typescript": "3.3.4000",
-        "vscode": "1.1.30",
-        "tslint": "5.14.0",
-        "@types/node": "^8.10.25",
-        "@types/mocha": "^2.2.42"
-    },
-    "dependencies": {
-        "azure-devops-node-api": "^7.0.0"
+    "views": {
+      "azure-boards-container": [
+        {
+          "id": "work-items",
+          "name": "My work items"
+        }
+      ]
     }
+  },
+  "scripts": {
+    "vscode:prepublish": "npm run compile",
+    "compile": "tsc -p ./",
+    "watch": "tsc -watch -p ./",
+    "postinstall": "node ./node_modules/vscode/bin/install",
+    "test": "npm run compile && node ./node_modules/vscode/bin/test"
+  },
+  "devDependencies": {
+    "typescript": "3.3.4000",
+    "vscode": "1.1.30",
+    "tslint": "5.14.0",
+    "@types/node": "^8.10.25",
+    "@types/mocha": "^2.2.42"
+  },
+  "dependencies": {
+    "azure-devops-node-api": "^7.0.0"
+  }
 }
