+++ src/extension.ts
@@ -2,7 +2,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import { WorkItemTreeNodeProvider } from "./workitems/workitem";
+import { WorkItemTreeNodeProvider } from "./workitems/workitem.tree";
 import { GitExtension } from "./externals/git";
 
 // this method is called when your extension is activated
